refactor(wallet): consolidate fiat formatter in utils
