+++ client/src/components/pages/Store/Wallet/Transactions/Payment/components/PaymentSuccessContent.jsx
@@ -12,9 +12,8 @@ import { useTranslations } from "next-intl";
 import { useCurrency } from "@/components/hooks/useCurrency";
 import { CopyButton } from "@/components/shared/CopyButton";
 
-import { formatSats } from "../../../utils/formatters";
+import { formatFiat, formatSats } from "../../../utils/formatters";
 import { useSatsToFiatEstimate } from "../hooks/useSatsToFiatEstimate";
-import { formatFiat } from "../utils/formatFiat";
 
 export function PaymentSuccessContent({
   isOpen,
