+++ client/src/components/pages/Store/Wallet/Transactions/Payment/components/PendingPaymentContent.jsx
@@ -13,9 +13,8 @@ import { useTranslations } from "next-intl";
 
 import { useCurrency } from "@/components/hooks/useCurrency";
 
-import { formatSats } from "../../../utils/formatters";
+import { formatFiat as formatFiatValue, formatSats } from "../../../utils/formatters";
 import { usePaymentAmountInput } from "../hooks/usePaymentAmountInput";
-import { formatFiat as formatFiatValue } from "../utils/formatFiat";
 
 import { ZeroAmountPaymentFields } from "./ZeroAmountPaymentFields";
