+++ client/src/components/pages/Store/Wallet/Transactions/Payment/index.js
@@ -1,7 +1,6 @@
 export { PaymentTab } from "./components/PaymentTab";
 export { PaymentConfirmModal } from "./components/PaymentConfirmModal";
 export { PaymentSuccessContent } from "./components/PaymentSuccessContent";
-export { formatFiat } from "./utils/formatFiat";
 export { getPaymentErrorDescription, PAYMENT_ERROR_TRANSLATIONS } from "./utils/paymentErrors";
 export { usePaymentAmountInput } from "./hooks/usePaymentAmountInput";
 export { useSatsToFiatEstimate } from "./hooks/useSatsToFiatEstimate";
