--- client/src/components/pages/Store/Wallet/Transactions/Payment/utils/formatFiat.js
@@ -1,14 +1,0 @@
-export function formatFiat({
-  value,
-  currencyAcronym,
-  locale,
-}) {
-  try {
-    return new Intl.NumberFormat(locale || "en-US", {
-      style: "currency",
-      currency: currencyAcronym || "USD",
-    }).format(value);
-  } catch {
-    return `${currencyAcronym || "USD"} ${value.toFixed(2)}`;
-  }
-}
