+++ client/src/components/pages/Store/Wallet/utils/__tests__/formatters.test.js
@@ -1,4 +1,4 @@
-import { formatSats } from "../formatters";
+import { formatFiat, formatSats } from "../formatters";
 
 describe("formatSats", () => {
   it("formats numbers with thousand separators", () => {
@@ -29,5 +29,23 @@ describe("formatSats", () => {
 
   it("handles very large numbers", () => {
     expect(formatSats(1000000000)).toBe("1,000,000,000");
+  });
+});
+
+describe("formatFiat", () => {
+  it("formats fiat values using currency and locale", () => {
+    expect(formatFiat({
+      value: 12.5,
+      currencyAcronym: "USD",
+      locale: "en-US",
+    })).toBe("$12.50");
+  });
+
+  it("falls back when Intl currency formatting fails", () => {
+    expect(formatFiat({
+      value: 12.5,
+      currencyAcronym: "INVALID",
+      locale: "en-US",
+    })).toBe("INVALID 12.50");
   });
 });
